Document sync (#20)
Closes #10

Now errors can be shown even if the file isn't saved to disk. Also makes
the checking run in a separate thread which gets cancelled when there
are changes. Overall the experience is a lot nicer now.

Also makes `DocumentURI` be a distinct to stop issues popping up

Currently doesn't work with find usages/definition. But seeing as that
doesn't really work anyways, we aren't losing much.

Basic implementation, can be improved with better caching in future