Better action API (#27)
Closes #13 

Mainly just moving around functions, but is the start of a simplier API
for code actions. Based on editing nodes directly instead of text.

Will need a more advanced `PNode` that has access to the parent. Will
look into using a custom AST that supports this (Remember seeing some
kind of custom AST thing in the Nim codebase)