Slightly better goto definition (#51)
I found the old goto definition didn't work a lot (In fairness, it turns
out I forgot to enable it) and would usually fail since Nim's IC doesn't
really work. Instead moved to a system that uses the symbols parsed in
the current document. While this won't work cross document or in local
scopes it does do it a bit more of the time and can be improved