Add Nix flake (#54)
Packaging this as a flake so its easier to install.

Used my own setup instead of using the upstream buildNimblePackage since
this uses an FOD instead of requiring a specific lock. Might break it
out into its own repo so I can use it to package other packages I have